It's been another one of those days.. things may be hopelessly broken now.
The main theme here is that I'm trying to get rid of pathnames as inode identifiers.
#define PATHNAME_INODE will provide the old behavior.

The repercussions of this are numerous.  The major thing that I didn't get to was changing the way the 
client handles filename data.  I removed the 4k static char array and replaced it with a  
a variable length data pointer and counter at the end of the struct.  Filenames should go 
here, but should only be sent on open.  Hence the fid should be returned to the client with the reply and 
the client should continue to use that as the id.  

This also means that the analog namespace will come into play (mapsvc/zestFidLink.c).

I left things in a bad state, it doesn't compile but it's not too bad.  Only a few things need to be fixed 
but I really need to stop working so I'll do it later tonight or tomorrow.
p



git-svn-id: svn+ssh://frodo/cluster/svn/projects@732 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda