Added another arg to zthr_init (void *)arg which can be passed into the thread main struct. More header file madness (something is still broken in zestBlockDesc.h) 	#include <sys/types.h> @include/zestReadWrite.h
Added some compatiblity stuff to minimize the changes to lustre rpc
Example:
(zestLock.h)	 
+#define spinlock_t zest_spinlock_t
+#define SPIN_LOCK_UNLOCKED LOCK_INITIALIZER
+
+#define spin_lock_init(l) LOCK_INIT(l)
+
 #define freelock(l)                 

Added #define LOCK_PREFIX "" to include/zestAtomic.h which eliminates the need for 
asm/alternative.h.  LOCK_PREFIX  is set to "" for UP but not for SMP - this is probably not 
a long term solution.

zestThread.h - attempting to separate the thread structure from the zestion so that it may 
by used by the rpc threads.  As you can see more work needs to be done here to prevent the 
linking of out-of-context header files.  ie zestrpc doesn't care about zestSuperBlock.h




git-svn-id: svn+ssh://frodo/cluster/svn/projects@1065 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda