deploy the new custom format string API for logging, so we never have to make changes to zestion/log.c again!
how to use: set the environmental variable ZEST_LOG_FORMAT
to a string with the following printf(3)-like specifiers:

  %n	thread name (or "<pid>" if unavailable in e.g. zestFormat)
  %L	logging level number (2 for error, 4 notice, 6 info, etc.)
  %l	line number in file where error occurred (__LINE__)
  %s	gettimeofday tv_sec
  %u	gettimeofday tv_usec
  %f	filename where error occurred (__FILE__)
  %F	function name where error occurred (__func__)
  %h	hostname

you should be able to use modifiers such as %06u just like
you would for printf(3), and they are passed verbatim.

the default format is "[%s:%06u %n:%f:%l]", which closely
matches what was being used before.

in the future, we can even easily make the format adjustable
via zestiondctl to change on the fly.



git-svn-id: svn+ssh://frodo/cluster/svn/projects@1271 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda