Here's much reqoeking of zinodes and zchunks.. I haven't tested this but it's going in the right direction.  There are tons of sanity checks within the inode state machine, hopefully that will make debugging easy.
TODO - there's an outstanding issue with nseen as used by the free list.. it needs to 
be replaces with total..

TODO - zeils maintain inode pointers which may be invalidated since there's no 
guarantee that they won't be reaped - unless they've processed a write.  Currently 
there's no way to detect this but since we're maintaining a refcnt of 1 (when the zeil 
does its inode lookup) it shouldn't be a problem.. but we will not be able to reap 
inodes until refcnt is 0.




git-svn-id: svn+ssh://frodo/cluster/svn/projects@1583 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda