allow client RAID set paritygroup size to not have to be a multiple or factor of the # of iothreads!
this is accomplished by adding a probability factor
to the ingestgroup which the iothread uses when deciding
where to pull work from.  entries are added multiple times
in the iothread's zig list to increase (or decrease) the
chances (relative to other entries) of him pulling client
buffers from them.

known optimization that could be made here (with the current
implementation): find the lowest common demoninator among parity
group sizes and reduce the pointer array size



git-svn-id: svn+ssh://frodo/cluster/svn/projects@2241 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda