use PFL log!
this is perhaps the toughest PFL conversion since it touches
everything and changes the control interface

this is perhaps buggy.  however, i have a more complete diff
that i will be double checking and making sure relevant pieces
are put in, but if you still run into problems, let me know.



git-svn-id: svn+ssh://frodo/cluster/svn/projects@2791 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda