add a struct which mimics  fuse_dirent (not in use right now but we may need it later).
set FCMH_FCOO_CLOSING if the fcoo ref counts get to 0.

remove fidc_fcm_setattr from openrpc, we'll just depend on the previous getattr.
to fix this the server needs to fill in the attr struct contained in the reply.

slash2fuse_openref_update has to be wrapped with fcmh lock.

slash2fuse_readdir - just hand the readdir buf (from the rpc) right to fuse.





git-svn-id: svn+ssh://frodo/cluster/svn/projects@4798 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda