Here's the result of this week's adventure into inode cache stabilization.  Lingering problems may still exist but it seems to be working much better than before.
This diff changes the way inodes are reaped from the clean list back to the free 
list.  There are several important modifications:

1. The hash bucket is locked during the entire lookup process.  Conversly, 
the reclaim mechanism locks the hash bucket to prevent threads from discovering 
inodes which are about to be freed.  There are some nasty race conditions which 
can result from such a practice so the reclaim mechanism will yield if he can't 
lock the hash bucket. 

2. zinode_freelist_avail_check() works now.  It was badly broken before.

3. zinocache_reclaim() takes a lock on the clean list and iterates over the list 
searching for candidates.  Previously it would pull items off the list then 
inspect them - this was racy.  Inodes selected for reaping are set to ZINODE_FREEING
and their lock is released.  Hence zinocache_lookup() will now skip any ino entries
which are set to ZINODE_FREEING.




The fio test I used to break and test the inode cache was something like this:

group 8peReadWrite {
  files_per_dir = 128;
  tree_depth    = 5;
  tree_width    = 3;
  pes           = 8;
  test_freq     = 0; 
  block_freq    = 0;
  path          = /zzest/pauln/tmp;
  output_path   = /home/pauln/;
  filename      = 8peaasaaaaaaaadaacREAD;
  file_size     = 2m;
  block_size    = 128k;
  thrash_lock   = yes;
  samedir       = yes;
  samefile      = no;
  intersperse   = no;
  seekoff       = no;
  fsync_block   = no;
  verify        = yes;
  barrier       = yes;
  time_block    = no;
  block_barrier = no;
  time_barrier  = no;
  iterations    = 1;
  debug_conf    = no;
  debug_block    = no;
  debug_memory    = no;
  debug_buffer    = no;
  debug_output    = no;
  debug_dtree     = no;
  debug_barrier   = no;
  debug_iofunc    = no; 
                
  iotests (
        WriteEmUp [create:openwr:write:close]   
        statem  [stat]
  )     
}






git-svn-id: svn+ssh://frodo/cluster/svn/projects@5384 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda