revise zigoff code.
the old code maintained a zigoffv value in each inode representing, for
each file, the position that various workers should begin sending the
chunks they receive into the zig queues.

this value has been removed because small I/O done to many files
essentially obviates the usefulness of storing this value alongside
a file.  what needs to happen is application of this value across files
and even clients connections.

so now we maintain a global atomic_t that is bumped by a single
increment of the parity size in use for a parity group, and this value
is scaled to each new parity size to properly maintain offsets.



git-svn-id: svn+ssh://frodo/cluster/svn/projects@5496 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda