add a new thread to the bmap flush system for reaping completed rpcs.
change spinlock(&bmpce... to  BMPCE_LOCK(bmpce)

fix a bug in bmap_biorq_del() where the opcnt was being dropped 
prematurely.  Also in bmap_biorq_del(), remove references to BMAP_DIRTY2LRU.
We no longer need that since we do minor -> major structure locking order
for the bmaps and the bmapFlushQ.  this greatly simplifies the code.

Use a flag, BIORQ_DESTROY, to denote that a biorq is going away.

Only the bmap_flush thread may remove bmaps from the bmapFlushQ so therefore
we must wait for that unlinking operation to complete prior to releasing
a bmap.

Add a BIORQ_FLUSHRDY state to the biorq to prevent the flush thr from taking
a biorq from the page cache too early (writes only).  This state allows the 
biorq to be present on in the pending queue while being ignored by the flush
thread.

BMAP_CLI_FLUSHPROC state to denote that dirty bmaps have already been scheduled.

Decrease the timeout age for pages, this helped to shake out lots of timing
related bugs.

Increase MAX_OUTSTANDING_RPCS for the hell of it.

Try to be smarter about how many rpcs we schedule from the flush thr.

add bmap_flush_bmpce_check_sched_locked() which will help us ignore biorq's
which have scheduled pages.  Setting the same page inflight twice causes 
problems for the flush engine.

Add bmpc_wake_reaper() so that fuse threads dont spin in and out of the reaper
code.  If no pages are freed in the reap call we go to sleep and are woken
when a page is added to the lru.

Redo the LRU logic so that pages can be reclaimed more efficiently


git-svn-id: svn+ssh://frodo/cluster/svn/projects@9497 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda