remove mds_bmap_loadvalid() calls in the replication manager.
always load the bmap, needed to work around the tricky behavior
of waiting for the mount_slash bmap flush thread to eventually
have the ION send CRC updates to the MDS which will only then
"initialize" the bmap crcstates, at that point allowing replication
enqueue requests to proceed.

instead, to detect holes, when enqueuing, if no ACTIVE replicas
exist for the specified bmaps, skip.  this won't present an error
to the user, although it is easy to change it to do so.


git-svn-id: svn+ssh://frodo/cluster/svn/projects@9591 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda