in setattr, obtain fid's from mds if not present in the cache.  I was seeing EINVAL from setattr ops while untarring a kernel and running find.  This was on a client with a limit of 2048 fcmh's.  The lowered number caused dir fcmh's to be flushed prematurely.
This patch also tries to properly deal with dircache init and cleanup by 
doing a better job at detecting unintialized dircache info's.  The issue is that
sometimes at fcmh_ctor runtime, the attrs of the fid are not known.  The 
result is that we don't know whether to init the fci as a file or dir


git-svn-id: svn+ssh://frodo/cluster/svn/projects@12095 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda