introduce use-after-free operation when PFL_DEBUG > 1 (PFL_DEBUG is set to 1 by default).
so, if you want to use it, you have to do:

  $ DEBUG=2 make build

"linger" mode is implemented with a LRU with entries placed on one end
on free(3) and eventually pulled off on alloc(3) when the configured
list size is full.

also, move handling of realloc(p, 0) to the macros; its much easier.


git-svn-id: svn+ssh://frodo/cluster/svn/projects@14072 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda