first bits for improved error handling and recovery during the I/O process.
the default behavior is to wait for I/O servers to come online to
successfully service READ and WRITE RPCs.

if a client specifies O_NONBLOCK (which we need to check for updates via
fcntl(2), somehow...) then return EIO immediately.  note: there are still
parts needed to implement this, this is only a first step.

the O_NONBLOCK checks are consolidated at a single place,
msl_offline_retry(), so it is easy to change the determination method.


git-svn-id: svn+ssh://frodo/cluster/svn/projects@15645 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda