there is no way to detect if you have a read lock.
although you can partially infer, sometimes dead lock happens or EAGAIN
is returned for no reason.  really, the behavior seems undefined.

try this if you don't believe me:

	pthread_rwlock_unlock(&rw)
	pthread_rwlock_unlock(&rw)
	pthread_rwlock_rdlock(rw)

so resurrect the protection layer around pthread_rwlock and manage who
is doing what ourselves.

this might be expensive (because of dynarray) but i doubt it's much
worse than all the lock mangling we were doing before this.  and the
logic flows much more simply.  and it is more portable to other systems
(like BSD) whose rwlock implementations are even less forgiving than
linux (if that is possible).


git-svn-id: svn+ssh://frodo/cluster/svn/projects@16289 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda