A couple fixes here.
1) Simple implementation of client-side enforced write ordering.  The idea is that the client does not allow the same
page on the wire twice.  This is in response to observations of large writes being sent first but applied to the sliod 
backfs after newer, smaller writes were issued by the same client.

2) Create a blocking version of msl_bmap_lease_tryext() which is called when the lease is a bit 'too' old to 
rely on a non-blocking call.  Now when an old lease is encountered, we can guarantee a successful renewal before
issuing the I/O.

3) If biorq_destroy is called from msl_io() it would assert because biorq_destroy() assumes  biorq's had been 
on the wire.  msl_io calls to biorq_destroy() are really a short-circuit tripped when the I/O rpc could not be 
issued.  In this patch we change the logic so that failed RBW biorq's can issue  biorq_destroy()  without asserting
on the wire conditions



git-svn-id: svn+ssh://frodo/cluster/svn/projects@17203 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda