remove odtable from the replication engine, entirely.
it has been the source of so much woes and stress it really really is
not worth eeking to try to get it right.  odtable is made for very
simple, straightforward operations and is not malleable for the behavior
upsch requires.

sure, we may lose some replication or other activity.  but:

  (1) we are losing much more now by not being able to utilize the
      engine at all, because of the buggyness with odtable

  (2) in the worst case, we lose the entire table; oh well.
      persistent replications will kick on next access and otherwise
      users will monitor their file attributes and requeue.

  (3) the sqlite database is way more reliable anyway, since it sits
      outside of ZFS and relies on fsync(2), and we have not encountered
      a single problem with its consistency yet.

the new system works with a nonce value stored in the db.  unless we
retrieve the per-uptime nonce of the current system from the db for the
replication activity in question, the activity is reset to be queued.
if it is current, then it may be processed.

------------------------------------------------------------------------

also, with the new batch RPC API, reduce partial reclaim to two small
functions (!!)


git-svn-id: svn+ssh://frodo/cluster/svn/projects@21852 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda