workaround an unfortunate situation:
if a peer is connecting to us twice, something is awry.  the lustre
behavior sends an EOF to the new connection letting them know that a
prior connection is still open.

we have run into deployments where sockets are getting closed on only
one end, so the peer thinks a connection is open when it really isn't
(relative to the source node).  so, we just close all sockets when we
see this so state can re-initialize from square one.


git-svn-id: svn+ssh://frodo/cluster/svn/projects@24695 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda