please see https://github.com/pscedu/pfl/commit/a3b2752bb005725f15b4097e79d2df248295ecfa for why i am reverting the last change to the nbset reaper.
this is the source of the recent problems with daemons not being
able to reconnect to each other.

i really thought we were taking a more conservative approach to changes
instead of thought based approaches.  things are delicate.  test suites
need to incorporate failure modes because tolerating failure is one of
the premier features of successful distributed systems.